requirements.txt
pyodbc
pyarrow
boto3
pyiceberg[s3fs,hive]
psycopg2-binary
trino
pydantic-settings